[artf37282] : ConfigureScreenLocking(74) - askForPasswordDelay not set to 0
After looking at item 1. the screen saver options it looks like that is a stonix bug code says to set it to 0 but it is
not. Will submit that to the tracker for the next release of stonix.

The options that stonix is to set are:
defaults write com.apple.screensaver askForPassword -int 1
defaults write com.apple.screensaver askForPasswordDelay -int 0