[artf38049]: Updated SystemAccounting unit test
This UT was failing on Debian and Ubuntu due to apt's issues with
executing commands when it has a stale cache, and proxy issues. Fixed
both problems.
The rpm setup was also changed to use vars for repeated paths, and to
remove a logic error where the file was opened for writing.