[artf37985]: Improvements to SHupdaterc.isrunning
SHupdaterc.isrunning was only looking for "running" in the output of
"service [svc] status" command, but on Debian 7, stopped services output
"not running", with a return code of 3. Added chk.returncode == 0 to
make isrunning compatible with Deb 7 (and chk.wait() to ensure "service"
finishes execution before attempting to read returncode).