[artf41007]: Bypass simpleRuleTest in DisableGUILogon UT
Due to the fact that this test does not perform the REMOVEX fix
(entirely removing X Windows from the system, which can't be undone), it
was always showing failure. The test has been adjusted so that it
specifically checks just the DISABLEX and LOCKDOWNX CIs. REMOVEX will
have to continue to be tested during triage.