[artf41991]: Added profiles.ini to mozPath in AuditFFUsage UT
Ubuntu is being very picky, and was still managing to fail the UT when
run remotely with /root/.mozilla/firefox as the path, although there
were still no profiles created. Added profiles.ini (which is created
when the first profile is created) to the path to ensure that the rule
only runs when there is at least one profile present.