[artf42011]: RootMailAlias will check if aliases file is a symlink
This rule (and its unit test) were causing problems because they would
always write directly to /etc/aliases, but on Macs, this should be a
symlink. Changed the code for locating the /etc/aliases file to check if
it is a symlink, and if so, to use the linked path instead.