[artf42602]: Adjusted AuditFirefoxUsage to use two paths
After finding that the most accurate method to check for the existence
of Firefox profiles was to check for .mozilla/firefox/profiles.ini, the
mozPath variable was changed to this path. However, the unit test was
designed with a directory in mind, and as such some things did not work
for a single file. Added the profiles.ini path as profilePath and
reinserted .mozilla/firefox as mozPath.